math-algorithm-book3.4節完了
