+++ math-algorithm-book/src/main/kotlin/chapter03/section02/Question015.kt
@@ -1,4 +1,4 @@
-package chapter03
+package chapter03.section02
 
 // 問題
 // https://atcoder.jp/contests/math-and-algorithm/tasks/math_and_algorithm_o
