+++ math-algorithm-book/src/main/kotlin/chapter03/section02/sampleCode002.kt
@@ -1,4 +1,4 @@
-package chapter03
+package chapter03.section02
 
 import kotlin.math.min
